cluster work, cleanup.

+++ Uravo/Serverroles/Cluster.pm
@@ -3,6 +3,7 @@ package Uravo::Serverroles::Cluster;
 use strict;
 
 use Uravo;
+use Uravo::Util;
 
 our $uravo = $Uravo::uravo;
 
@@ -52,15 +53,15 @@ sub getTypes {
 }
 
 sub _list {
-    my $params = shift;
+    my $params = Uravo::Util::clean_params(shift || {});
 
     $uravo ||= new Uravo;
     $uravo->log("Uravo::Serverroles::Cluster::_list()",9);
 
-    my $cluster_id = ($params->{cluster} || $params->{cluster_id});
-    my $type_id = ($params->{type} || $params->{type_id});
-    my $silo_id = ($params->{silo} || $params->{silo_id});
-    my $bu_id = ($params->{bu} || $params->{bu_id});
+    my $cluster_id = $params->{cluster_id};
+    my $type_id = $params->{type_id};
+    my $silo_id = $params->{silo_id};
+    my $bu_id = $params->{bu_id};
 
     my $list = $uravo->getCache("Cluster:_list:$cluster_id:$type_id:$silo_id:$bu_id");
     return keys %$list if ($list);
@@ -87,15 +88,16 @@ sub _list {
 }
 
 sub getSilo {
-    my $self = shift || return;
+    my $self = shift || die;
 
+    $uravo ||= new Uravo;
     return $uravo->getSilo($self->get("silo_id"));
 }
 
 
 sub set {
-    my $self        = shift || return;
-    my $field       = shift || return;
+    my $self        = shift || die;
+    my $field       = shift || die;
     my $value       = shift;
 
     return $self->update($field, $value);
@@ -107,6 +109,8 @@ sub update {
     my $field           = shift || return;
     my $value           = shift;
     my $changelog       = shift;
+
+    $uravo ||= new Uravo;
 
     if (!$changelog && ref($value) eq 'HASH' && ref($field) eq 'HASH') {
         $changelog = $value;
@@ -122,11 +126,30 @@ sub update {
     }
 
     if ($field eq 'netblock_id') {
+        foreach my $netblock_id (@$value) {
+            my $netblock = $uravo->getNetblock($netblock_id) || die "invalid netblock '$netblock_id'";
+            if ($netblock->get('silo_id') != $self->get('silo_id')) { die "$netblock_id is not in the same silo."; }
+        }
         $uravo->{db}->do("DELETE FROM cluster_netblock WHERE cluster_id=?", undef, ($self->id())) || die($uravo->{db}->errstr);
         foreach my $netblock_id (@$value) {
             $uravo->{db}->do("INSERT INTO cluster_netblock (cluster_id, netblock_id, create_date) VALUES (?,?,NOW())", undef, ($self->id(), $netblock_id));
         }
 
+        return $changelog;
+    }
+    elsif ($field eq 'silo_default') {
+        $value = MinorImpact::isTrue($value)?1:0;
+        die if ($self->get('silo_default') == $value);
+        die "you can't unset the default cluster, you have to set a new cluster as default." unless ($value);
+
+        my $silo_id = $self->get("silo_id");
+        $uravo->{db}->do("UPDATE cluster SET silo_default = 0 WHERE silo_id=?", undef, ($self->silo_id())) || die($uravo->{db}->errstr);
+        $uravo->{db}->do("UPDATE cluster SET silo_default = 1 WHERE cluster_id=?", undef, ($self->id())) || die($uravo->{db}->errstr);
+        $self->{silo_default} = 1;
+
+        if ($changelog) {
+            $uravo->changelog({object_type=>$self->{object_type}, object_id=>$self->id(), field_name=>$field, old_value=>"0", new_value=>"1"},$changelog);
+        }
         return $changelog;
     }
 
@@ -168,6 +191,8 @@ sub link {
 sub changelog {
     my $self = shift || return;
 
+    $uravo ||= new Uravo;
+
     my $data = $uravo->{db}->selectall_arrayref("SELECT cl.id, cl.user, cl.ticket, cl.note, cl.create_date FROM changelog cl WHERE cl.object_id=? AND cl.object_type=? ORDER BY cl.create_date desc", {Slice=>{}}, ($self->id(), $self->type()));
     my @changelogs = ();
     foreach my $changelog (@$data) {
@@ -183,13 +208,19 @@ sub changelog {
 }
 
 sub add {
-    my $params  = shift || return;
+    my $params = Uravo::Util::clean_params(shift || {});
     my $changelog = shift;
 
-    my $cluster_id = $params->{cluster_id} || return;
-    my $silo_id = $params->{silo_id} || return;
+    $uravo ||= new Uravo;
 
-    $uravo->{db}->do("INSERT INTO cluster (cluster_id, silo_id, create_date) VALUES (?, ?, now())", undef, ($cluster_id, $silo_id)) || die ($uravo->{db}->errstr);
+    my $cluster_id = $params->{cluster_id} || return;
+    my $silo_id = $params->{silo_id} || die "you must specify a silo";
+    my $silo_default = MinorImpact::isTrue($params->{silo_default})?1:0;
+
+    my $data = $uravo->{db}->selectrow_hashref("SELECT COUNT(*) AS c FROM cluster WHERE silo_id=?", undef, ($silo_id)) || die ($uravo->{db}->errstr);
+    $silo_default = 1 unless ($data->{c});
+
+    $uravo->{db}->do("INSERT INTO cluster (cluster_id, silo_id, silo_default, create_date) VALUES (?, ?, ?, now())", undef, ($cluster_id, $silo_id, $silo_default)) || die ($uravo->{db}->errstr);
 
     if ($changelog) {
         $uravo->changelog({object_type=>'cluster', object_id=>$cluster_id, field_name=>'New cluster', new_value=>$cluster_id},$changelog);
@@ -199,14 +230,35 @@ sub add {
 }
 
 sub delete {
-    my $params = shift || return;
+    my $params = shift || die;
     my $changelog = shift || {};
 
-    my $cluster_id = $params->{cluster_id} || return;
+    $params = Uravo::Util::clean_params($params);
 
     $uravo ||= new Uravo;
-    foreach my $server ($uravo->getServers({cluster=>$cluster_id})) {
-        $server->update({cluster_id=>'unknown'},$changelog);
+
+    my $cluster_id = $params->{cluster_id} || die "no cluster_id specified";
+
+    if ($cluster_id eq "unknown") {
+        die "It is forbidden to kill the unknown cluster.";
+    }
+
+    my $cluster = $uravo->getCluster($cluster_id) || die "cluster $cluster_id is invalid";
+    my $silo_id = $cluster->silo_id();
+      
+    my $data = $uravo->{db}->selectrow_hashref("SELECT cluster_id FROM cluster WHERE silo_id=? AND silo_default = 1", undef, ($silo_id)) || die ($uravo->{db}->errstr);
+    my $default_cluster_id = $data->{cluster_id};
+
+    if ($cluster_id eq $default_cluster_id) {
+        die "can't delete the default cluster";
+    }
+    elsif ($default_cluster_id eq "") {
+        die "no default cluster for $silo_id";
+    }
+    else {
+        foreach my $server ($uravo->getServers({silo_id=>$silo_id, cluster=>$cluster_id})) {
+            $server->update({cluster_id=>$default_cluster_id},$changelog);
+        }
     }
 
     $uravo->{db}->do("delete from changelog, changelog_detail using changelog inner join changelog_detail where changelog.object_id=? and changelog.object_type='cluster' and changelog.id=changelog_detail.changelog_id", undef, ($cluster_id)) || die($uravo->{db}->errstr);
@@ -228,22 +280,33 @@ sub data {
 sub getNetblocks {
     my $self = shift || return;
     my $params = shift || {};
+    my $local_params = MinorImpact::cloneHash($params);
 
-    my $local_params = {};
-    foreach my $key (keys %$params) {
-        $local_params->{$key} = $params->{$key};
-    }
+    $uravo ||= new Uravo;
+
     $local_params->{cluster_id} = $self->id();
 
     return $uravo->getNetblocks($local_params);
 }
 
+sub info {
+    my $self = shift || die;
+    my $params = shift || {};
+
+    my $output = "";
+    $output .= "name:" . $self->id() . "\n";
+    $output .= "silo_id:" . $self->getSilo()->id() . "\n";
+
+    print $output;
+}
+
 # Misc info functions.
+sub get     { my ($self, $name) = @_; return $self->{$name}; }
 sub id		{ my ($self) = @_; return $self->{cluster_id}; }
 sub name	{ my ($self) = @_; return $self->{name} || $self->id(); }
-sub version { my ($self) = @_; return $self->{version}; }
+sub silo_id { my ($self) = @_; return $self->{silo_id}; }
 sub type    { my ($self) = @_; return $self->{object_type}; }
-sub get     { my ($self, $name) = @_; return $self->{$name}; }
+sub version { my ($self) = @_; return $self->{version}; }
 
 
 1;
